everything at this point

+++ backend/db/migrations/20230611212040-create-review.js
@@ -26,7 +26,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(3,1)
       },
       createdAt: {
         allowNull: false,
