+++ backend/db/models/review.js
@@ -20,7 +20,7 @@ module.exports = (sequelize, DataTypes) => {
     spotId: DataTypes.INTEGER,
     userId: DataTypes.INTEGER,
     review: DataTypes.STRING,
-    stars: DataTypes.INTEGER
+    stars: DataTypes.DECIMAL(3,1)
   }, {
     sequelize,
     modelName: 'Review',
